+++ jdklog-core/src/main/java/org/jdkstack/jdklog/logging/core/handler/Execute.java
@@ -1,0 +1,5 @@
+package org.jdkstack.jdklog.logging.core.handler;
+
+public interface Execute {
+  //
+}
